+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Merged branches and closed issues
 * Interpolating 2D analysis, Optimization analysis (!180)
 * Allow for unit-aware printing of floats and other values with no error (!167, #193)
 * Quantities of interest saving to JSON now supports more types, including ufloat (!164, #152)
+* Added new utility class quantify.utilities.inspect_utils
 
 0.4.0 (2021-05-10)
 ------------------